Disable all animations except "static" ("solid_color_anim")
I can't seem to figure out how to sync up the animations with the
correct OpenRGB names for them, so I'm just giving up and disabling all
but the "static" one. The "Direct" animation is added automatically
regardless.

Basically, when you cycle through the various animations, OpenRGB was
associating them with the wrong ones. Ideally, I could enable whichever
ones I wish and OpenRGB would match them with the right name, but that
ain't happening, so I cut out all but "Static" and "Direct", and now it
works.

I've also experimented with some success by tweaking OpenRGB, but I'm
not wanting to maintain a fork of OpenRGB just for this, so I'll leave
it alone for now.