Add a mouse layer to use WASD and arrow keys as mouse and knob as scroll
This can be toggled by clicking the encoder button. It can also be
activated when Caps Lock is held. Once active, you can tap Caps Lock or
click the encoder button to switch back.

When active, the WASD and arrow keys move the mouse, and E becomes right
click and both Q and Space become left click.

Additionally, all of the mentioned keys turn teal on this layer.